+++ shopify_integration/shopify_integration/doctype/shopify_payout/shopify_payout.py
@@ -8,9 +8,9 @@ import frappe
 from frappe.model.document import Document
 from frappe.utils import cint, flt
 
-from shopify_integration.connector import create_sales_return, get_tax_account_head
+from shopify_integration.invoices import create_sales_return
 from shopify_integration.shopify_integration.doctype.shopify_log.shopify_log import make_shopify_log
-from shopify_integration.utils import get_accounting_entry
+from shopify_integration.utils import get_accounting_entry, get_tax_account_head
 
 
 class ShopifyPayout(Document):
